requirements.txt
requests
PySimpleGUI
praw
yfinance
python-dotenv
pandas
numpy
nltk
transformers
torch
